validacion (Nota y campos obligatorios)
para la nota, y campos al llenar el form1

+++ CapaEntidad/CENota.cs
@@ -5,7 +5,7 @@
         public string IdUsuario { get; set; }       // ID del usuario
         public string CodigoMateria { get; set; }   // Código de la materia
         public string IdCategoria { get; set; }     // ID de la categoría de la nota
-        public decimal Calificacion { get; set; }   // Nota obtenida
+        public double Calificacion { get; set; }   // Nota obtenida
         public string Comentario { get; set; }      // Comentario adicional
     }
 }
